@@ -24,7 +24,7 @@
     "@babel/preset-env": "^7.15.8",
     "@babel/preset-react": "^7.16.0",
     "@types/loadable__component": "5.13.4",
-    "@types/react": "^17.0.35",
+    "@types/react": "18",
     "babel-loader": "^8.2.2",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^9.0.1",
@@ -56,9 +56,10 @@
   },
   "dependencies": {
     "@loadable/component": "^5.15.0",
-    "@sqlframes/repl-app": "^0.13.0",
+    "@sqlframes/repl-app": "^2.5.4",
+    "@types/react-dom": "18",
     "date-fns": "^2.26.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "18",
+    "react-dom": "18"
   }
 }
