@@ -1,14 +1,12 @@
-import React, { Component, useRef, useEffect } from "react";
+import React, { Component, useMemo } from "react";
 import "./App.css";
 import { REPLReact } from "./REPLReact";
-import { DataFrameComponent } from "./DataFrameComponent";
 
 // this is an example to bundle SQL Frames directly into the app
 
 // Destructuring assignment of the exported namespaces 
-import { sqlframes, repl } from '@sqlframes/repl-app';
-// to ensure the css is also bundled into the application css bundle
-import '@sqlframes/repl-app/styles';
+import { sqlframes } from '@sqlframes/repl-app';
+import SQLFramesComponent from "./SQLFramesComponent";
 
 // All the top level classes and objects exported in the namespaces
 const { DataFrame, SQL, Time, View } = sqlframes;
@@ -20,25 +18,20 @@ class App extends Component{
 }
 
 const ChartApp = () => {
-	const chartRef = useRef(null);
-	const df = DataFrame.fromURL('https://earthquake.usgs.gov/earthquakes/feed/v1.0/summary/all_month.csv');
-	const { groupBy, where: { eq }, agg: { count } } = SQL;
-	const chart = () => {
+	const [df,chart] = useMemo(() => {
+		const df = DataFrame.fromURL('https://earthquake.usgs.gov/earthquakes/feed/v1.0/summary/all_month.csv');
+		const { groupBy, where: { eq }, agg: { count } } = SQL;
 		const gdf = df.gdf(groupBy('type'));
-		return gdf.chart.bar('type','Count')
+		return [df,gdf.chart.bar('type','Count')
 					.resize(600,400)
-					.svg();
-	};
-
-	useEffect(() =>{
-		View.render(chartRef.current,chart());
-	});
+					.svg()];
+	},[]);
 
 	return (
-		<div>
-			<div ref={chartRef}></div>
-			<DataFrameComponent df={df}/>
-		</div>
+		<>
+			<SQLFramesComponent value={chart}/>
+			<SQLFramesComponent value={df}/>
+		</>
 	)
 }
 
